First signs of networking life
- Add object packet dif'ing to only send delta snapshots to the
  server. Diffs between successive frames are sent

+++ engine/src/renderer/parallax.cpp
@@ -101,11 +101,11 @@ void Parallax::render(Renderer* renderer, const SDL_Rect& clip)
     auto transformed_dst = layer.bbox;
     if (is_foreground) {
       // Things that are closer move faster 
-      transformed_dst.x -= cx * (1.0 + layer.z_index / 100.0);
-      transformed_dst.y = -layer.surface->h + GAME_HEIGHT - (clip.y * (1.0 - layer.z_index / 100.0));
+      transformed_dst.x -= static_cast<int32_t>(cx * (1.0 + layer.z_index / 100.0));
+      transformed_dst.y = static_cast<int32_t>(-layer.surface->h + GAME_HEIGHT - (clip.y * (1.0 - layer.z_index / 100.0)));
     } else {
-      transformed_dst.x -= cx * (1.0 - (layer.z_index / 100.0));
-      transformed_dst.y = -layer.surface->h + GAME_HEIGHT - (clip.y * (layer.z_index / 100.0));
+      transformed_dst.x -= static_cast<int32_t>(cx * (1.0 - (layer.z_index / 100.0)));
+      transformed_dst.y = static_cast<int32_t>(-layer.surface->h + GAME_HEIGHT - (clip.y * (layer.z_index / 100.0)));
     }
 
     SDL_RenderCopyEx(renderer->sdl.renderer, layer.texture.get(), &layer.bbox, &transformed_dst, 0, nullptr, SDL_FLIP_NONE);
